Moved code between the backwards compatibility and the new delegate object headers
