+++ src/app/clusters/device-energy-management-server/app_config_dependent_sources.gni
@@ -13,6 +13,7 @@
 # limitations under the License.
 app_config_dependent_sources = [
   "DeviceEnergyManagementCluster.cpp",
+  "DeviceEnergyManagementDelegate.h",
   "DeviceEnergyManagementTestEventTriggerHandler.h",
   "device-energy-management-server.h",
 ]
